Fix ordering of clamp arguments
Even though mix is defined as (x, y, a) [1], clamp is defined as (a, x,
y) [2] where x is the lowerbound and y is the upperbound. This breaks on
Linux, which enforces the WebGL argument ordering and is likely working
on the other platforms where it patches the argument ordering on the
fly.

[1]: https://www.khronos.org/registry/OpenGL-Refpages/gl4/html/mix.xhtml
[2]: https://www.khronos.org/registry/OpenGL-Refpages/gl4/html/clamp.xhtml