Add Support for using the Space key to click cells
PROXX so far has a lot of ease-of-use shortcuts to support desktop users. Things like using the F key to change the mode, using the arrow keys to navigate the board, and using the Enter key to click cells. However, it is pretty awkward for users who want to play the game using only a keyboard to do so; for example, on a U.S. keyboard, the Enter key is miles away from the F key, therefore requiring the user to either stretch their hand for a prolonged period of time, or position their fingers on both the f and 8 keys (which is a really unusual position for most keyboard users, especially the ones who use the Home-row typing strategy). The situation is a bit less awkward for users with a U.K keyboard due to the fact that the hashtag key is right next to the Enter key; however, that requires the user to place their left hand on the Enter and hashtag keys, while placing their right hand on the arrow keys, therefore causing both hands to overlap. This issue gets even worse on laptops because the arrow keys are significantly closer to the Enter key.

The solution, therefore, is to allow the Space key to be used to click on cells. This will please both the users who use the Home-row typing strategy (by placing their left hand normally (i.e. on the F key) and using their thumb to click on space) and the users who are used to traditional gaming keybinds (they will need to, however, get used to placing their hand on the F key instead of WASD, but that's definitely much easier to get used to than the F&8 position).