Add WASD support for left-handed users
PROXX currently supports using the arrow keys for navigating the board. However, keyboard users whose dominant hand is their left one might find it difficult to use the arrow keys when compared to the WASD keys. This also applies, to a lesser extent, to users who are used to the WASD keys as a way to move around in games. Adding support for these keys will provide fast and reliable movements for keyboard users who are trying to get a new high score, or just regular users who want to play for a prolonged period of time without having their non-dominant hand hurt.