Add support for changing flag mode using the M key
As an unfortunate result of adding support for WASD, PROXX needs to add an easy way for keyboard users to use their right hand for changing the flag mode and for clicking cells. Adding support for the Space key fixes the latter problem. For the former, though, the current keybinds don't work on the U.S. keyboard because the F key is hard to reach with one's right hand (and even if a user decides to do so, it'll cause overlap between that hand and their left hand (which is on the WASD keys). This isn't a problem on the U.K. keyboard, however, since the hashtag key is right next to the Enter key, but supporting the U.S. users is still crucial. Using the M key to change the *M*ode makes a lot of sense, and it would allow a seamless experience for users.